+++ thread.h
@@ -4,10 +4,12 @@
 
 #include "prime.h"
 #include <stdio.h>
+#include <stdlib.h>
 #include <pthread.h>
 #include <unistd.h>
-#include <time.h>
+#include <sys/time.h>
 
-int createThreads(int number);
+
+int createThreads(long number);
 
 #endif
